+++ Cargo.toml
@@ -30,7 +30,7 @@ members = [
   "examples/raw_upload",
   "examples/pastebin",
   "examples/state",
-  "examples/state_nontrivial",
+  "examples/managed_queue",
   "examples/uuid",
   "examples/session",
   "examples/raw_sqlite",
